systems/compass/requirements.txt
streamlit==1.32.0
openai==1.12.0
pyyaml==6.0.1
python-dotenv==1.0.0
watchdog==3.0.0
supabase==2.3.4
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-mock==3.12.0
